Added year scale
Now use start of next unit for tracking end, and start of this unit
for tracking beginning.

I.e, checking earliest start date will compare the start of the month
when using monthly scale, and the start of the year when using yearly
scale. For end date it will use start of next month.